# docs/doc.json/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/regent": {
            "get": {
                "description": "get all Regents",
                "tags": [
                    "Regents"
                ],
                "summary": "Get all Regents",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/main.Regent"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/regent/{idregent}": {
            "get": {
                "description": "get regent by idregent",
                "tags": [
                    "Regents"
                ],
                "summary": "Get one regent",
                "parameters": [
                    {
                        "type": "string",
                        "description": "idregent",
                        "name": "idregent",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.Regent"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "main.Regent": {
            "type": "object",
            "properties": {
                "area": {
                    "type": "string"
                },
                "capital": {
                    "type": "string"
                },
                "idregent": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        }
    }
}